Rework the layout of the user interface
Initially the layout used dashboardPage from the package shinydashboard,
but a single-page layout seemed more natural for the project that a
dashboard with tabs.  The new layout uses fluidPage from the shiny
package.  A CSS string inserted into the head tag tweaks font sizes in
order to highlight the structure of the layout.